Added grid.hpp and movement.hpp
The grid class is meant for use with the computer vision.
It is a representation of the board as an 8x8 grid rather than the 8x4 playable-columns-only representation the board class uses.

The movement classes SensorGroup, Cart, and Arm are skeletons for a high level CBC motor and sensor interface.
Not much more development can happen there until the actual robot is built.